Fix an overflow in the delta time calculation.  A double as incorrectly converted to a long.

+++ usr.sbin/dntpd/convert.c
@@ -31,7 +31,7 @@
  * OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  * 
- * $DragonFly: src/usr.sbin/dntpd/convert.c,v 1.1 2005/04/24 02:36:50 dillon Exp $
+ * $DragonFly: src/usr.sbin/dntpd/convert.c,v 1.2 2005/04/24 06:23:56 dillon Exp $
  */
 
 #include "defs.h"
@@ -83,7 +83,7 @@ tv_add_micro(struct timeval *tvp, long usec)
 double
 tv_delta_micro(struct timeval *tv1, struct timeval *tv2)
 {
-    long usec;
+    double usec;
 
     usec = (double)(tv2->tv_sec - tv1->tv_sec) * 1000000.0 +
 		(double)(tv2->tv_usec - tv1->tv_usec);
